% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.spatial.index_function.R
\name{calc.spatial.index}
\alias{calc.spatial.index}
\title{Calculate spatial indices for points in 2D space}
\usage{
calc.spatial.index(xcor, ycor, res = 1)
}
\arguments{
\item{xcor}{X-coordinate of the point}

\item{ycor}{Y-coordinate of the point}

\item{res}{side length of one spatial subunit}
}
\value{
spatial index for the input point
}
\description{
Function that returns spatial indices (plot numbers) for given coordinate pairs.
E.g. for hectares res = 100 m
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{index}
\keyword{number}
\keyword{plot}
\keyword{spatial}

