% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colmx2ras_function.R
\name{colmx2ras}
\alias{colmx2ras}
\title{Convert bycol-matrix to raster}
\usage{
colmx2ras(mx)
}
\arguments{
\item{mx}{Input matrix}
}
\value{
Raster object
}
\description{
Convert a matrix with origin in upper left corner and with cell order from
top to bottom (bycol) to a raster with origin in lower left corner
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{bycol}
\keyword{conversion}
\keyword{coordinate}
\keyword{matrix}
\keyword{origin}
\keyword{raster}
\keyword{transposition}

