\name{LidarProcessoR-package}
\alias{LidarProcessoR-package}
\alias{LidarProcessoR}
\docType{package}
\title{
\packageTitle{LidarProcessoR}
}
\description{
\packageDescription{LidarProcessoR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LidarProcessoR}
\packageIndices{LidarProcessoR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LidarProcessoR}

Maintainer: \packageMaintainer{LidarProcessoR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
